Feat: 유튜브를 통해서 동영상을 다운받을 수 있도록 만듬

+++ externalSystem/app/_global/workdir/WorkDirManager.py
@@ -6,8 +6,9 @@ from ..logger import CustomLogger
 from ..logger import CustomLoggerType
 
 class WorkDirManager:
-    def __init__(self):
+    def __init__(self, isAfterClear:bool=True):
         self.workDirPath = "./workDirs/" + str(uuid.uuid4()) + "/"
+        self.isAfterClear = isAfterClear
           
     def __enter__(self):
         CustomLogger.debug(CustomLoggerType.EFFECT, "Try to make workdir", "<workDirPath: {}>".format(self.workDirPath))
@@ -15,5 +16,6 @@ class WorkDirManager:
         return lambda subPath="" : self.workDirPath + subPath
       
     def __exit__(self, exc_type, exc_value, exc_traceback):
-        CustomLogger.debug(CustomLoggerType.EFFECT, "Try to remove workdir", "<workDirPath: {}>".format(self.workDirPath))
-        shutil.rmtree(self.workDirPath)
+        if self.isAfterClear :
+            CustomLogger.debug(CustomLoggerType.EFFECT, "Try to remove workdir", "<workDirPath: {}>".format(self.workDirPath))
+            shutil.rmtree(self.workDirPath)
